Add NamedEntity
